Reorder header in models page
