+++ frontend/page_models/REstimateSeriesView.tsx
@@ -23,9 +23,13 @@ export function REstimateSeriesView(props: ModelViewProps) {
     height?: number;
   }>({});
 
+  let [plotlyHtmlElement, setPlotlyHtmlElement] = React.useState<
+    Plotly.PlotlyHTMLElement
+  >();
+
   // the rescale callback, will be used as an callback ref
   const rescale = React.useCallback((node: HTMLElement | null) => {
-    if (!node) {
+    if (!node || !plotlyHtmlElement) {
       setDimensions({ width, height });
       return;
     }
@@ -33,11 +37,14 @@ export function REstimateSeriesView(props: ModelViewProps) {
     const idealHeight = window.innerHeight * 0.7;
     const maxWidth = idealHeight * MAX_CHART_WIDTH_RATIO;
     const maxHeight = idealWidth * MAX_CHART_HEIGHT_RATIO;
-
     setDimensions({
       node,
       width: Math.max(Math.min(idealWidth, maxWidth), MIN_CHART_SIZE),
       height: Math.max(Math.min(idealHeight, maxHeight), MIN_CHART_SIZE),
+    });
+    Plotly.relayout(plotlyHtmlElement, {
+      "xaxis.autorange": true,
+      "yaxis.autorange": true,
     });
   }, []);
 
@@ -51,7 +58,7 @@ export function REstimateSeriesView(props: ModelViewProps) {
   }, [node]);
 
   // create a plotly config for the plot
-  let { config, hook } = React.useMemo(
+  let { config } = React.useMemo(
     () =>
       makeConfig(() => {
         if (!region) {
@@ -69,27 +76,12 @@ export function REstimateSeriesView(props: ModelViewProps) {
     []
   );
 
-  let makeResponsive = () => {
-    if (!node) return;
-
-    (node.querySelector(
-      "#r_estimate_dataviz .svg-container"
-    ) as HTMLElement).removeAttribute("style");
-
-    let mainSvg = node.querySelector(
-      "#r_estimate_dataviz .main-svg"
-    ) as SVGSVGElement;
-    mainSvg.removeAttribute("width");
-    mainSvg.removeAttribute("height");
-    mainSvg.setAttribute("viewBox", `0 0 ${width} ${height}`);
-  };
-
   // create a layout and customize
   let layout = makeLayout();
 
   layout.width = width;
   layout.height = height;
-
+  layout.autosize = true; // set autosize to rescale
   layout.margin!.r = 20;
   layout.xaxis!.type = "date";
   layout.yaxis!.title = "R estimation";
@@ -121,17 +113,18 @@ export function REstimateSeriesView(props: ModelViewProps) {
       <h5 className="mitigation-heading">
         Effective reproduction number estimate:
       </h5>
-      <div id="r_estimate_dataviz" ref={rescale}>
-        <Plot
-          style={{}}
-          data={data}
-          layout={layout}
-          config={config as any}
-          onUpdate={(_: any, gd: Plotly.PlotlyHTMLElement) => makeResponsive()}
-          onInitialized={(_: any, gd: Plotly.PlotlyHTMLElement) => {
-            hook(gd);
-          }}
-        />
+      <div>
+        <div id="r_estimate_dataviz" ref={rescale}>
+          <Plot
+            style={{}}
+            data={data}
+            layout={layout}
+            config={config as any}
+            onInitialized={(_: any, gd: Plotly.PlotlyHTMLElement) => {
+              setPlotlyHtmlElement(gd);
+            }}
+          />
+        </div>
       </div>
     </>
   );
